requirements.txt
fastapi
uvicorn[standard]
requests
redis
python-dotenv
openai
apscheduler
confluent-kafka
confluent-kafka[avro]
httpx